+++ classes/Url.php
@@ -37,8 +37,17 @@ class Url
 
     /**
      * @return UriInterface
+     * @deprecated Use the psr method instead
      */
     public function getUri(): UriInterface
+    {
+        return $this->url;
+    }
+
+    /**
+     * @return UriInterface
+     */
+    public function psr(): UriInterface
     {
         return $this->url;
     }